feat: Location picker for event creation menu You can now select locations for new events
